Fixes nested usage of `listKeys` mocks #2829
