+++ data/secret-template.json
@@ -1,0 +1,190 @@
+{
+  "Microsoft.Storage/storageAccounts": {
+    "listKeys": {
+      "keys": [
+        {
+          "keyName": "key1",
+          "permissions": "Full",
+          "value": "{{Secret}}"
+        },
+        {
+          "keyName": "key2",
+          "permissions": "Full",
+          "value": "{{Secret}}"
+        }
+      ]
+    }
+  },
+  "Microsoft.DocumentDB/databaseAccounts": {
+    "listKeys": {
+      "primaryMasterKey": "{{Secret}}",
+      "secondaryMasterKey": "{{Secret}}",
+      "primaryReadonlyMasterKey": "{{Secret}}",
+      "secondaryReadonlyMasterKey": "{{Secret}}"
+    }
+  },
+  "Microsoft.ApiManagement/service": {
+    "listKeys": {
+      "primary": "{{Secret}}",
+      "secondary": "{{Secret}}"
+    }
+  },
+  "Microsoft.AppConfiguration/configurationStores": {
+    "listKeys": {
+      "value": [
+        {
+          "id": "439AD01B4BE67DB1",
+          "name": "Primary",
+          "value": "{{Secret}}",
+          "connectionString": "{{Secret}}",
+          "lastModified": "2018-04-24T16:30:54+00:00",
+          "readOnly": false
+        },
+        {
+          "id": "CB45E100456857B9",
+          "name": "Secondary",
+          "value": "{{Secret}}",
+          "connectionString": "{{Secret}}",
+          "lastModified": "2018-04-24T16:30:54+00:00",
+          "readOnly": false
+        },
+        {
+          "id": "B3AC55B7E71431A9",
+          "name": "Primary Read Only",
+          "value": "{{Secret}}",
+          "connectionString": "{{Secret}}",
+          "lastModified": "2018-04-24T16:30:54+00:00",
+          "readOnly": true
+        },
+        {
+          "id": "E2AF6A9A89DCC177",
+          "name": "Secondary Read Only",
+          "value": "{{Secret}}",
+          "connectionString": "{{Secret}}",
+          "lastModified": "2018-04-24T16:30:54+00:00",
+          "readOnly": true
+        }
+      ]
+    }
+  },
+  "Microsoft.Automation/automationAccounts": {
+    "listKeys": {
+      "keys": [
+        {
+          "KeyName": "Primary",
+          "Permissions": "Full",
+          "Value": "{{Secret}}"
+        },
+        {
+          "KeyName": "Secondary",
+          "Permissions": "Full",
+          "Value": "{{Secret}}"
+        }
+      ]
+    }
+  },
+  "Microsoft.Batch/batchAccounts": {
+    "listKeys": {
+      "accountName": "sampleacct",
+      "primary": "{{Secret}}",
+      "secondary": "{{Secret}}"
+    }
+  },
+  "Microsoft.Cache/redis": {
+    "listKeys": {
+      "primaryKey": "{{Secret}}",
+      "secondaryKey": "{{Secret}}"
+    }
+  },
+  "Microsoft.CognitiveServices/accounts": {
+    "listKeys": {
+      "key1": "{{Secret}}",
+      "key2": "{{Secret}}"
+    }
+  },
+  "Microsoft.Devices/IotHubs": {
+    "listKeys": {
+      "value": [
+        {
+          "keyName": "iothubowner",
+          "primaryKey": "{{Secret}}",
+          "secondaryKey": "{{Secret}}",
+          "rights": "RegistryWrite, ServiceConnect, DeviceConnect"
+        },
+        {
+          "keyName": "service",
+          "primaryKey": "{{Secret}}",
+          "secondaryKey": "{{Secret}}",
+          "rights": "ServiceConnect"
+        },
+        {
+          "keyName": "device",
+          "primaryKey": "{{Secret}}",
+          "secondaryKey": "{{Secret}}",
+          "rights": "DeviceConnect"
+        },
+        {
+          "keyName": "registryRead",
+          "primaryKey": "{{Secret}}",
+          "secondaryKey": "{{Secret}}",
+          "rights": "RegistryRead"
+        },
+        {
+          "keyName": "registryReadWrite",
+          "primaryKey": "{{Secret}}",
+          "secondaryKey": "{{Secret}}",
+          "rights": "RegistryWrite"
+        }
+      ]
+    }
+  },
+  "Microsoft.NotificationHubs/namespaces/notificationHubs/authorizationRules": {
+    "listKeys": {
+      "primaryConnectionString": "{{Secret}}",
+      "secondaryConnectionString": "{{Secret}}",
+      "primaryKey": "{{Secret}}",
+      "secondaryKey": "{{Secret}}",
+      "keyName": "sdk-AuthRules-5800"
+    }
+  },
+  "Microsoft.NotificationHubs/namespaces/authorizationRules": {
+    "listKeys": {
+      "primaryConnectionString": "{{Secret}}",
+      "secondaryConnectionString": "{{Secret}}",
+      "primaryKey": "{{Secret}}",
+      "secondaryKey": "{{Secret}}",
+      "keyName": "RootManageSharedAccessKey"
+    }
+  },
+  "Microsoft.MachineLearningServices/workspaces": {
+    "listKeys": {
+      "userStorageKey": "{{Secret}}",
+      "userStorageResourceId": "/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/rg-test/providers/Microsoft.Storage/storageAccounts/sttest",
+      "appInsightsInstrumentationKey": "{{Secret}}",
+      "containerRegistryCredentials": {
+        "location": null,
+        "username": "{{SecretUsername}}",
+        "passwords": [
+          {
+            "name": "password",
+            "value": "{{Secret}}"
+          },
+          {
+            "name": "password2",
+            "value": "{{Secret}}"
+          }
+        ]
+      },
+      "notebookAccessKeys": {
+        "primaryAccessKey": "{{Secret}}",
+        "secondaryAccessKey": "{{Secret}}"
+      }
+    }
+  },
+  "Microsoft.MachineLearning/webServices": {
+    "listKeys": {
+      "primary": "{{Secret}}",
+      "secondary": "{{Secret}}"
+    }
+  }
+}
